# storage/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nutrients API",
        "description": "REST API for managing ingredients and their nutritional data.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/ingredients/{ingredient}/nutrients": {
            "get": {
                "tags": [
                    "IngredientNutrients"
                ],
                "summary": "List nutrients attached to an ingredient",
                "operationId": "c4110bc2112c2eb8304574f751d787f9",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nutrient list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Nutrient"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "IngredientNutrients"
                ],
                "summary": "Detach multiple nutrients from an ingredient",
                "operationId": "1405d8b366e16faedc5009397c32e4f8",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nutrient_ids"
                                ],
                                "properties": {
                                    "nutrient_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Detached"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
        "/api/ingredients/{ingredient}/nutrients/attach": {
            "post": {
                "tags": [
                    "IngredientNutrients"
                ],
                "summary": "Attach a nutrient to an ingredient",
                "operationId": "3d008aefee89333d24eb8baaac2144f4",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nutrient_id"
                                ],
                                "properties": {
                                    "nutrient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5,
                                        "nullable": true
                                    },
                                    "amount_unit_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated nutrient list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Nutrient"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
        "/api/ingredients/{ingredient}/nutrients/{nutrient}": {
            "put": {
                "tags": [
                    "IngredientNutrients"
                ],
                "summary": "Update pivot data for a nutrient on an ingredient",
                "operationId": "1a75e9333eef06f9eebefacb6329a5a4",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nutrient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "nullable": true
                                    },
                                    "amount_unit_id": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated nutrient list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Nutrient"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "IngredientNutrients"
                ],
                "summary": "Detach a nutrient from an ingredient",
                "operationId": "f6719f327d40027f24c94e66ab1907e5",
                "parameters": [
                    {
                        "name": "ingredient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nutrient",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Detached"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "frontend": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Nutrient": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "external_id": {
                        "type": "string",
                        "example": "1004",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Vitamin C"
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "vitamin-c"
                    },
                    "iu_to_canonical_factor": {
                        "type": "number",
                        "format": "float",
                        "example": null,
                        "nullable": true
                    },
                    "is_label_standard": {
                        "type": "boolean",
                        "example": true
                    },
                    "display_order": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "sync_status": {
                        "type": "string",
                        "example": "synced"
                    },
                    "source": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Source"
                            }
                        ],
                        "nullable": true
                    },
                    "canonical_unit": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Unit"
                            }
                        ],
                        "nullable": true
                    },
                    "parent": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Nutrient"
                            }
                        ],
                        "nullable": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Nutrient"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NutrientTag"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "NutrientTag": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Fat-soluble"
                    },
                    "slug": {
                        "type": "string",
                        "example": "fat-soluble"
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Source": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "USDA FoodData Central"
                    },
                    "slug": {
                        "type": "string",
                        "example": "usda-fooddata-central"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://fdc.nal.usda.gov",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Unit": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Gram"
                    },
                    "abbreviation": {
                        "type": "string",
                        "example": "g"
                    },
                    "type": {
                        "type": "string",
                        "example": "mass",
                        "nullable": true
                    },
                    "to_base_factor": {
                        "type": "number",
                        "format": "float",
                        "example": 1,
                        "nullable": true
                    },
                    "base_unit_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "frontend": {
                "type": "http",
                "description": "Bearer token obtained from POST /api/auth/login. Also requires X-Refresh-Token, X-Application-Name, and X-Client-Url headers on every request.",
                "scheme": "bearer"
            }
        }
    },
...
}